% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachmentscoring.R
\name{attachment_tool}
\alias{attachment_tool}
\title{Attahchment Anxiety and Avoidance Measurement Tool}
\usage{
attachment_tool(df, firstq)
}
\arguments{
\item{df}{dataframe}

\item{firstq}{first col in the series of attachment style cols}
}
\value{
df with new cols for attachment anxiety and attachment avoidance scores
}
\description{
The purpose of this function is to score participants responses to the 12-item
Wei et. al Attachment Measure and provide their attachment anxiety and attachment avoidance scores.
}
\examples{
attachment_tool(mydata, "attachmentQ1")
}
