% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investments.so.far.R
\name{investments.so.far}
\alias{investments.so.far}
\title{Investments So Far}
\usage{
investments.so.far(df, firstq, numinvest)
}
\arguments{
\item{df}{the name of your dataframe}

\item{firstq}{the name of your first weekly investment col (going in chronoloigical order from left to right)}

\item{numinvest}{the number of weekly investment cols you have in your dataframe}
}
\value{
a invests.so.far col that will tell the researcher how many investments each participant has
made into their relationship at any given week during the study.
}
\description{
The purpose of this function is to determine how many investments participants have made
so far at any given week in the Canada Pair Project
}
\examples{
investments.so.far(mydata, met_friends, 5)
}
