% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.variables.R
\name{lag.variables}
\alias{lag.variables}
\title{Create Lag Variables for Cross Lag Panel Model
Description: Helper function to create variables for the cross lag panel model}
\usage{
lag.variables(df, x, y, z, a)
}
\arguments{
\item{df}{dataframe}

\item{x}{first variable of interest}

\item{y}{second variable of interest}

\item{z}{third variable of interest}

\item{a}{fourth variable of interest}
}
\value{
four variables in which the data is extracted from the week before the current date
}
\description{
Create Lag Variables for Cross Lag Panel Model
Description: Helper function to create variables for the cross lag panel model
}
\examples{
lag.variables(mydata, attachment, attraction, mood, investment)
}
